[account settings] updated account settings
created components based off of different tabs under account settings
