+++ src/app/(profile)/account-manage/SocialMediaTab.tsx
@@ -1,0 +1,10 @@
+"use client";
+import { Card, CardBody } from "@nextui-org/react";
+
+export default function SocialMediaTab() {
+  return (
+    <Card className="w-full sm:w-[65%]">
+      <CardBody>this is a test</CardBody>
+    </Card>
+  );
+}
